requirements.txt
pandas>=2.2.3
openpyxl>=3.1.2
python-dateutil>=2.9.0
pytz>=2025.2
# Note: For Windows systems only:
pywin32==306
msal>=1.22.0
msgraph-core>=0.2.2
azure-identity>=1.12.0
requests>=2.31.0 